Updated Foundation Code [Functionality]:  Added a day object to the ItemModel Struct
Added a way to identify each task so it can be called and grouped for later use.  The value for the property is instanced in the weekdayMenuController.swift.

Additional notes: Still haven't found a way to filter and group the tasks by day, tests show the value is being stored correctly.